+++ src/device.rs
@@ -1,6 +1,10 @@
 //! Supported devices.
 
-use crate::{crates, probe::Probe, utils::ser_to_string};
+use crate::{
+    crates,
+    probe::{Log, Probe},
+    utils::ser_to_string,
+};
 use anyhow::Result;
 use serde::{Deserialize, Serialize};
 use std::io::Write;
@@ -438,6 +442,44 @@ impl Device {
             | Self::Stm32L4S9 => &[Probe::Bmp, Probe::Openocd],
             Self::Nrf52810 | Self::Nrf52811 | Self::Nrf52832 | Self::Nrf52840 => &[Probe::Openocd],
             Self::Nrf9160 => &[Probe::Jlink],
+        }
+    }
+
+    /// Returns the list of supported debug loggers.
+    pub fn logs(&self) -> &[Log] {
+        match self {
+            Self::Stm32F100
+            | Self::Stm32F101
+            | Self::Stm32F102
+            | Self::Stm32F103
+            | Self::Stm32F107
+            | Self::Stm32F401
+            | Self::Stm32F405
+            | Self::Stm32F407
+            | Self::Stm32F410
+            | Self::Stm32F411
+            | Self::Stm32F412
+            | Self::Stm32F413
+            | Self::Stm32F427
+            | Self::Stm32F429
+            | Self::Stm32F446
+            | Self::Stm32F469
+            | Self::Stm32L4X1
+            | Self::Stm32L4X2
+            | Self::Stm32L4X3
+            | Self::Stm32L4X5
+            | Self::Stm32L4X6
+            | Self::Stm32L4R5
+            | Self::Stm32L4R7
+            | Self::Stm32L4R9
+            | Self::Stm32L4S5
+            | Self::Stm32L4S7
+            | Self::Stm32L4S9
+            | Self::Nrf52810
+            | Self::Nrf52811
+            | Self::Nrf52832
+            | Self::Nrf52840 => &[Log::SwoProbe, Log::SwoSerial],
+            Self::Nrf9160 => &[Log::DsoSerial],
         }
     }