+++ src/new.rs
@@ -4,7 +4,7 @@ use crate::{
     cli::NewCmd,
     crates,
     device::Device,
-    probe::{Probe, ProbeLog},
+    probe::{Log, Probe},
     templates::Registry,
     utils::run_command,
 };
@@ -20,7 +20,7 @@ use termcolor::{Color, ColorSpec, StandardStream, WriteColor};
 impl NewCmd {
     /// Runs the `drone new` command.
     pub fn run(&self, shell: &mut StandardStream) -> Result<()> {
-        let Self { path, device, flash_size, ram_size, probe, probe_log, name, toolchain } = self;
+        let Self { path, device, flash_size, ram_size, probe, log, name, toolchain } = self;
         let registry = Registry::new()?;
         let name = name.as_ref().map(String::as_str).map_or_else(
             || {
@@ -55,7 +55,7 @@ impl NewCmd {
             }
         }
         cargo_toml(path, &name, &device, &registry, shell)?;
-        drone_toml(path, &device, *flash_size, *ram_size, &probe, &probe_log, &registry, shell)?;
+        drone_toml(path, &device, *flash_size, *ram_size, &probe, &log, &registry, shell)?;
         justfile(path, &device, &registry, shell)?;
         rust_toolchain(path, &toolchain, &registry, shell)?;
         cargo_config(path, &registry, shell)?;
@@ -165,16 +165,15 @@ fn drone_toml(
     flash_size: u32,
     ram_size: u32,
     probe: &Option<Probe>,
-    probe_log: &ProbeLog,
+    log: &Option<Log>,
     registry: &Registry<'_>,
     shell: &mut StandardStream,
 ) -> Result<()> {
     let path = path.join("Drone.toml");
     let mut file = File::create(&path)?;
     let probe = probe.as_ref().unwrap_or_else(|| device.probes().first().unwrap());
-    file.write_all(
-        registry.new_drone_toml(device, flash_size, ram_size, probe, probe_log)?.as_bytes(),
-    )?;
+    let log = log.as_ref().unwrap_or_else(|| device.logs().first().unwrap());
+    file.write_all(registry.new_drone_toml(device, flash_size, ram_size, probe, log)?.as_bytes())?;
     print_created(shell, "Drone.toml")
 }