+++ src/probe/openocd.rs
@@ -1,6 +1,6 @@
 //! OpenOCD interface.
 
-use super::{run_gdb_client, run_gdb_server, rustc_substitute_path, setup_uart_endpoint};
+use super::{run_gdb_client, run_gdb_server, rustc_substitute_path, setup_serial_endpoint};
 use crate::{
     cli::{ProbeFlashCmd, ProbeGdbCmd, ProbeLogCmd, ProbeResetCmd},
     log,
@@ -13,7 +13,7 @@ use signal_hook::iterator::Signals;
 use std::{fs::File, path::PathBuf, process::Command, thread};
 use tempfile::tempdir_in;
 
-/// OpenOCD `drone probe reset` command.
+/// `drone probe reset` command with OpenOCD.
 #[allow(missing_docs)]
 pub struct ResetCmd<'a> {
     pub cmd: &'a ProbeResetCmd,
@@ -35,7 +35,7 @@ impl ResetCmd<'_> {
     }
 }
 
-/// OpenOCD `drone probe flash` command.
+/// `drone probe flash` command with OpenOCD.
 #[allow(missing_docs)]
 pub struct FlashCmd<'a> {
     pub cmd: &'a ProbeFlashCmd,
@@ -57,7 +57,7 @@ impl FlashCmd<'_> {
     }
 }
 
-/// OpenOCD `drone probe gdb` command.
+/// `drone probe gdb` command with OpenOCD.
 #[allow(missing_docs)]
 pub struct GdbCmd<'a> {
     pub cmd: &'a ProbeGdbCmd,
@@ -92,31 +92,31 @@ impl GdbCmd<'_> {
     }
 }
 
-/// OpenOCD `drone probe log` SWO command.
+/// `drone probe log` command with OpenOCD and SWO.
 #[allow(missing_docs)]
 pub struct LogSwoCmd<'a> {
     pub cmd: &'a ProbeLogCmd,
     pub signals: Signals,
     pub registry: Registry<'a>,
     pub config: &'a config::Config,
-    pub config_probe_swo: &'a config::ProbeSwo,
     pub config_probe_openocd: &'a config::ProbeOpenocd,
+    pub config_probe_swo: &'a config::ProbeSwo,
 }
 
 impl LogSwoCmd<'_> {
     /// Runs the command.
     pub fn run(self) -> Result<()> {
-        let Self { cmd, signals, registry, config, config_probe_swo, config_probe_openocd } = self;
+        let Self { cmd, signals, registry, config, config_probe_openocd, config_probe_swo } = self;
         let ProbeLogCmd { reset, outputs } = cmd;
 
         let mut _pipe_dir = None;
         let ports = outputs.iter().flat_map(|output| output.ports.iter().copied()).collect();
         let input;
         let commands;
-        if let Some(uart_endpoint) = &config_probe_swo.uart_endpoint {
-            setup_uart_endpoint(&signals, uart_endpoint, config_probe_swo.baud_rate)?;
-            exhaust_fifo(uart_endpoint)?;
-            input = uart_endpoint.into();
+        if let Some(serial_endpoint) = &config_probe_swo.serial_endpoint {
+            setup_serial_endpoint(&signals, serial_endpoint, config_probe_swo.baud_rate)?;
+            exhaust_fifo(serial_endpoint)?;
+            input = serial_endpoint.into();
             commands = registry.openocd_swo(config, &ports, *reset, None)?
         } else {
             let pipe_dir = tempdir_in(temp_dir())?;